Added responsedto

+++ src/main/java/nl/inholland/bankingapplication/controllers/UserAccountController.java
@@ -2,9 +2,7 @@ package nl.inholland.bankingapplication.controllers;
 
 import jakarta.persistence.EntityNotFoundException;
 import nl.inholland.bankingapplication.models.UserAccount;
-import nl.inholland.bankingapplication.models.dto.ExceptionDTO;
-import nl.inholland.bankingapplication.models.dto.UserAccountDTO;
-import nl.inholland.bankingapplication.models.dto.UserAccountUpdateDTO;
+import nl.inholland.bankingapplication.models.dto.*;
 import nl.inholland.bankingapplication.services.UserAccountService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -25,7 +23,7 @@ public class UserAccountController {
 
     @PreAuthorize("hasRole('ROLE_EMPLOYEE')")
     @GetMapping
-    public ResponseEntity getAllUserAccounts() {
+    public ResponseEntity<List<UserAccountResponseDTO>> getAllUserAccounts() {
         try{
             return ResponseEntity.ok(userAccountService.getAllUserAccounts());
         } catch (EntityNotFoundException e) {
@@ -35,7 +33,7 @@ public class UserAccountController {
 
     @PreAuthorize("hasRole('ROLE_EMPLOYEE')")
     @GetMapping("Exclude/{id}")
-    public ResponseEntity getAllUserAccountsExceptOne(@PathVariable Long id) {
+    public ResponseEntity<List<UserAccountResponseDTO>> getAllUserAccountsExceptOne(@PathVariable Long id) {
         try{
             return ResponseEntity.ok(userAccountService.getAllUserAccountsExceptOne(id));
         } catch (EntityNotFoundException e) {
@@ -45,7 +43,7 @@ public class UserAccountController {
 
     @PreAuthorize("hasRole('ROLE_EMPLOYEE')")
     @GetMapping("registered")
-    public ResponseEntity getAllRegisteredUserAccounts() {
+    public ResponseEntity<List<UserAccountResponseDTO>> getAllRegisteredUserAccounts() {
         try{
             return ResponseEntity.ok(userAccountService.getAllRegisteredUserAccounts());
         } catch (EntityNotFoundException e) {
@@ -75,7 +73,7 @@ public class UserAccountController {
 
 
     @PostMapping("register")
-    public ResponseEntity<UserAccount> addUserAccount(@RequestBody UserAccountDTO userAccountDTO) {
+    public ResponseEntity<UserAccountResponseDTO> addUserAccount(@RequestBody UserAccountDTO userAccountDTO) {
         try {
             return ResponseEntity.status(201).body(userAccountService.addUserAccount(userAccountDTO));
         } catch (EntityNotFoundException e) {
@@ -87,8 +85,14 @@ public class UserAccountController {
     @DeleteMapping("{id}")
     public ResponseEntity deleteUserAccount(@PathVariable Long id) {
         try {
-            userAccountService.deleteUserAccount(id);
-            return ResponseEntity.status(204).build();
+            UserAccount user = userAccountService.getUserAccountById(id);
+            if (user.getBankAccounts().size() > 0) {
+                return ResponseEntity.status(400).body("User with accounts cannot be deleted.");
+            }
+            else {
+                userAccountService.deleteUserAccount(id);
+                return ResponseEntity.status(204).build();
+            }
         } catch (EntityNotFoundException e) {
             return this.handleException(404, e);
         }
@@ -96,9 +100,9 @@ public class UserAccountController {
 
     @PreAuthorize("principal.username == @userAccountService.getUserAccountById(#id).username OR hasRole('ROLE_EMPLOYEE')")
     @PutMapping("{id}")
-    public ResponseEntity<UserAccount> updateUserAccount(@PathVariable Long id, @RequestBody UserAccountUpdateDTO userAccountDTO) {
+    public ResponseEntity<UserAccountResponseDTO> updateUserAccount(@PathVariable Long id, @RequestBody UserAccountUpdateDTO userAccountUpdateDTO) {
         try {
-            return ResponseEntity.status(200).body(userAccountService.updateUserAccount(id, userAccountDTO));
+            return ResponseEntity.status(200).body(userAccountService.updateUserAccount(id, userAccountUpdateDTO));
         } catch (EntityNotFoundException e) {
             return this.handleException(404, e);
         }
